Abstract helper functions to a helper file
